UnitsPerPx now includes upper axis limit
This way an axis spanning units [-10, +10] will result in a graph where the far left pixel column is -10 and the far right pixel column contains +10.

Previously a data point at 10 would appear one column to the right of the data area and would not be displayed on the graph.

This will improve the appearance of ticks and grid lines which will now fill the final column.