Backend.Resize(): useDelayedRendering argument
Resizing and rendering once is important for establishing plot dimensions.

Rendering is required because it measures string sizes required to determine the size of the axis labels and ticks, and determines the size of the data area.

Delayed rendering after resizing is good for mouse interactions, but a non-delayed render is required during initialization such as after Reset().

#1098