ScottPlot5 Plot: refactor Axis management
The primary goal here is to have the Plot module have a single object for each axis (not separate Axis and AxisView objects that have to be kept in sync).

Now that X and Y axes are stored in lists, the way is paved to support multi-axis plots.

This is an intermediate refactoring step. The API needs to be improved.