delete partial solution files
The main reason they were added is because building for Android was painfully slow. Now "rebuild all" isn't so bad, and users can manually unload the projects they do not wish to build. Also, working only in filtered solution files does not make it obvious when a code change breaks a project elsewhere in the solution.