Quality of life update for BarPlot with OADate x-axis
Created static class to provide const values for Bar widths when using a BarPlot with an OADateTime x-axis.  This aligns bar's in an adjacent fashion in scenarios where a bar represents a value collected over time. This can be done manually, this static class is just for ease of use.