Rendering: clear canvas at the start of each render
Previously this was not done, but the figure was being filled with an opaque background each time so previous renders were effectively erased. A problem arises when the figure background color is semitransparent, resulting in an effect where one can see multiple previous renders. Explicitly clearing the canvas at the start of each render ensures this does not happen and improves support for semitransparent backgrounds.