Plot: move mouse manipulations into Control.StandardActions
I'm trying to keep all the logic together for actions that controls call involving the mouse and keyboard.

This logic doesn't really belong in the Plot module itself because it's so specific to GUI actions.

It's a bit difficult to locate this logic as it is now though, so these functions should probably be broken into individual classes that inheritc IControlAction or something.