Adding hand drawn line styles to (most) plots
Added interface to query if the object has the LineStyle property. I usually like interfaces to be methods instead of holding a property, but this was an easy way to retrofit existing plottable classes.
Added SetLineStylePatterns method to the Styler class.
Added some unit tests to validate the line style pattern is being set properly.