Improve Annotations positioning
Stop using the magic number 4 and/or half the text height as an
approximate, hard coded margin.  Instead set a larger default OffsetX
and OffsetY for the Annotation (these two values are an effective margin
for Annotations). It is a more consistent margin, especially for larger
font sizes.  And, unlike the previous way of doing things, it can be
configured however the caller would like by adjusting the OffsetX and
OffsetY as they see fit.