AxisLine: Add property to allow exclusion from the legend
If an AxisLine (horizontal or vertical) has its Text property set it
would always be included in anyt legend - both the automatic or manually
built legends. The Text property is also used to display the text along
the appropriate axis for the line so it is a valid use case to want the
Text set but not add the line to the legend.

This change adds an ExcludeFromLegend property to the AxisLine base
class which defaults to false (no change to current default behavior).
Setting this property to true keeps the line out of the legend
regardless of the Text property being set or not.