`Generate.Consecutive` and friends for DateTimeOffset (#4863)
* Generate: DateTimeOffset functions

This includes most of the DateTime functions, but not those which are
deprecated, or those which take primitive parameters (because there's no
way for the compiler to distinguish a call to ConsecutiveDays(int year, int month,
int day) by return type alone).

I did not add overloads to the financial datagen functions which take
DateTimes, since it isn't clear to me how we should handle that
conversion. Any of assuming UTC, assuming local time, or taking an extra timezone/tzoffset parameter make sense to me. I think we should just let the user decide, it would be the most intuitive.

* Generate.RandomOHLCS(int, DateTime): Actually use the startDate passed in

* Add Generate.Consecutive sanity tests for DateTime(Offset)

* Update CHANGELOG.md

---------

Co-authored-by: Scott W Harden <swharden@gmail.com>